refactor(encoding): Phase 1e Stage B batch 8 — relocate SIMD rep-code probes onto impl HcMatcher
Move the five `for_each_repcode_candidate_with_reps*` functions to
`impl HcMatcher`:

  - `for_each_repcode_candidate_with_reps` — kernel dispatcher
  - `for_each_repcode_candidate_with_reps_neon` — AArch64 NEON
  - `for_each_repcode_candidate_with_reps_sse42` — x86 SSE4.2
  - `for_each_repcode_candidate_with_reps_avx2_bmi2` — x86 AVX2+BMI2
  - `for_each_repcode_candidate_with_reps_scalar` — fallback

Each now takes `&MatchTable` explicitly instead of reaching through
`self.table.X`. The kernel-specific variants keep their
`#[target_feature]` umbrellas so the per-rep `common_prefix_len_ptr`
call still inlines under the SIMD context, matching the donor
parity guarantee.

The `for_each_repcode_candidate_body!` macro is rewritten to consume
`$table` rather than `$self.table.X`, and gains `#[macro_export]` so
`hc::HcMatcher` can invoke it via `crate::for_each_repcode_candidate_body!`.
The `collect_optimal_candidates_initialized_body!` macro now
dispatches through `$self.hc.$for_each_rep(&$self.table, …)` to reach
the relocated methods from the BT side.

Test callers (`hc.for_each_repcode_candidate_with_reps(…)`) gain the
`&hc.table` arg via the same `<var>.hc.<method>(&<var>.table, …)`
pattern Stage A established for storage accessors.

Verification:
  - cargo build (default + --no-default-features) clean
  - cargo clippy --all-targets -- -D warnings clean
  - cargo nextest run -p structured-zstd --lib → 381 passed
